% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_sos.R
\name{calculate_sos}
\alias{calculate_sos}
\title{Calculate Strength of Schedule (SOS)}
\usage{
calculate_sos(ratings, schedule)
}
\arguments{
\item{ratings}{dataframe containing \code{team} and \code{rating}}

\item{schedule}{dataframe containing \code{team}, \code{opponent}, and \code{home_or_away}}
}
\value{
returns a dataframe with two columns - \code{team}, \code{sos}
}
\description{
Calculate strength of schedule using team ratings and schedule, created using
\code{calculate_team_ratings()} and \code{get_fixtures()}
}
\examples{
\dontrun{
remaining_schedule <-
  worldfootballR::fb_match_results(
    country = "ENG", gender = "M", season_end_year = 2024
    ) |>
  get_fixtures(remaining = TRUE)

team_ratings <-
  worldfootballR::fb_season_team_stats(
    country = "ENG", gender = "M", season_end_year = 2024,
    stat_type = "standard", tier = "1st"
    ) |>
  calculate_team_ratings()

calculate_sos(ratings = team_ratings, schedule = remaining_schedule)
}
}
