% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_team_ratings.R
\name{calculate_team_ratings}
\alias{calculate_team_ratings}
\title{Calculate team ratings}
\usage{
calculate_team_ratings(team_stats)
}
\arguments{
\item{team_stats}{dataframe containing FBref standard team stats (created
using \code{worldfootballR::fb_season_team_stats()} or following the same structure)}
}
\value{
returns a dataframe with four columns - \code{team}, \code{npxgd90}, \code{npgd90},
\code{rating}
}
\description{
Calculate team ratings from non-penalty goal difference per 90 and non-penalty
expected goal difference per 90.
}
\examples{
\dontrun{
team_stats <-
  worldfootballR::fb_season_team_stats(
    country = "ENG", gender = "M", season_end_year = 2024,
    stat_type = "standard", tier = "1st"
    )

calculate_team_ratings(team_stats)
}
}
