% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_fixtures.R
\name{get_fixtures}
\alias{get_fixtures}
\title{Get past or remaining league fixtures}
\usage{
get_fixtures(matches, remaining = TRUE)
}
\arguments{
\item{matches}{dataframe containing match results (created using
\code{worldfootballR::fb_match_results()} or following the same structure)}

\item{remaining}{get remaining fixtures (\code{TRUE}) or past fixtures (\code{FALSE})
(default = \code{TRUE})}
}
\value{
returns a dataframe with three columns - \code{team}, \code{home_or_away},
\code{opponent} - with each fixture appearing twice, once for the home team and
once for the away team
}
\description{
Returns the past or remaining league fixtures for every team in the league,
with each fixture shown twice (once for the home team and once for the away
team), from a match results dataframe created using
\code{worldfootballR::fb_match_results()} or following the same structure.
}
\examples{
\dontrun{
results <-
  worldfootballR::fb_match_results(
    country = "ENG", gender = "M", season_end_year = 2024
    )

get_fixtures(results, remaining = TRUE)
}
}
